+++ pytorch_lightning/lite/lite.py
@@ -26,7 +26,13 @@ from torch.utils.data import DataLoader, DistributedSampler, RandomSampler, Sequ
 
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.lite.wrappers import _LiteDataLoader, _LiteModule, _LiteOptimizer
-from pytorch_lightning.plugins import DDPSpawnPlugin, DeepSpeedPlugin, PLUGIN_INPUT, TPUSpawnPlugin, TrainingTypePlugin
+from pytorch_lightning.plugins import (
+    DDPSpawnStrategy,
+    DeepSpeedPlugin,
+    PLUGIN_INPUT,
+    TPUSpawnPlugin,
+    TrainingTypePlugin,
+)
 from pytorch_lightning.plugins.training_type.training_type_plugin import TBroadcast
 from pytorch_lightning.trainer.connectors.accelerator_connector import AcceleratorConnector
 from pytorch_lightning.utilities import _AcceleratorType, _StrategyType, move_data_to_device
@@ -310,7 +316,7 @@ class LightningLite(ABC):
         """
         if isinstance(obj, nn.Module):
             if self.device.type == "cuda":
-                # need to call this manually here again in case we spawned with DDPSpawnPlugin
+                # need to call this manually here again in case we spawned with DDPSpawnStrategy
                 # TODO: refactor to let plugin handle this cleanly
                 torch.cuda.set_device(self.device)
             return obj.to(self.device)
@@ -403,7 +409,7 @@ class LightningLite(ABC):
         # apply sharded context to prevent OOM
         run_method = partial(self._run_with_sharded_context, run_method)
 
-        if isinstance(self._strategy, DDPSpawnPlugin):
+        if isinstance(self._strategy, DDPSpawnStrategy):
             return self._strategy.spawn(run_method, *args, **kwargs)
         else:
             return run_method(*args, **kwargs)